+++ __tests__/controllers/engines/engine.view.test.js
@@ -1,84 +1,75 @@
+const { getCompilationConfigs, getBlueprint, getRuntimeMethods, getTemplate, buildTemplate, getFileString, render,
+  Template } = require('../../../controllers/engines/view')[Symbol.for('UNIT_TEST')];
+
+
 // mock
+const fs = require('fs');
+
+const mockLocals = { settings: {}, cache: false, _locals: {}, test: {} };
+const mockString = '<a>Hello World</a>';
+const mockCallback = (err, str) => {
+  if (err) return err;
+  return str;
+};
+
 jest.mock('fs', () => ({
   readFileSync: jest.fn(),
   readFile: jest.fn(),
 }));
-const fs = require('fs');
-
-
-// module
-const { getCompilationConfigs, getBlueprint, getRuntimeMethods,
-  getTemplate, buildTemplate, getFileString, render, Template,
-} = require('../../../controllers/engines/view')[Symbol.for('UNIT_TEST')];
+fs.readFileSync.mockReturnValue(mockString);
+fs.readFile.mockImplementation((file, option, cb) => cb(null, mockString));
 
 
 // test
-describe('Check the ENV', () => {
-  test('Should run in test mode', () => {
-    expect(process.env.NODE_ENV).toEqual('test');
-  });
-});
-
-
-describe('Bundle: engine.js', () => {
-  const mockLocals = { settings: {}, cache: false, _locals: {}, test: {} };
-  const mockContent = '<a>Hello World</a>';
-  fs.readFileSync.mockReturnValue(mockContent);
-  fs.readFile.mockImplementation((file, option, next) => next(null, mockContent));
-
-  test('Fn: getCompilationConfigs: Should generate doT configs on-the-fly', () => {
-    const result = getCompilationConfigs('a, b, c');
-    //
-    expect(result.varname).toBe('a, b, c');
+describe('Engines: View', () => {
+  test('Fn: getCompilationConfigs', () => {
+    const test = getCompilationConfigs('a, b, c');
+    expect(test.varname).toBe('a, b, c');
   });
 
-  test('Fn: getBlueprint: Should transpile Express.js meta into the blueprint of Template{object}', () => {
-    const result = getBlueprint(mockLocals);
-    //
-    expect(result).toHaveProperty('_fn');
-    expect(result).toHaveProperty('test');
-    expect(result).not.toHaveProperty('settings');
-    expect(result).not.toHaveProperty('cache');
-    expect(result).not.toHaveProperty('_locals');
+
+  test('Fn: getFileString', async () => {
+    const test = await Promise.all([getFileString('test', true), getFileString('test').then(str => str)]);
+    expect(test[0]).toEqual(mockString);
+    expect(test[1]).toEqual(mockString);
   });
 
-  test('Fn: getRuntimeMethods: Should get runtime template methods', () => {
-    const result = getRuntimeMethods({}, {}, '');
-    //
-    expect(result).toHaveProperty('useMarkdown');
-    expect(result).toHaveProperty('loadPartial');
+
+  test('Fn: getRuntimeMethods', () => {
+    const test = getRuntimeMethods({}, {}, '');
+    expect(test).toHaveProperty('useMarkdown');
+    expect(test).toHaveProperty('loadPartial');
   });
 
-  test('Fn: getTemplate: Should get a compiled Template{object}', async () => {
-    const result = [getTemplate('test', {}, true), await getTemplate('test', {})];
+
+  test('Fn: getBlueprint', () => {
     //
-    expect(result[0] instanceof Template).toBeTruthy();
-    expect(result[1] instanceof Template).toBeTruthy();
+    const test = getBlueprint(mockLocals);
+    expect(test).toHaveProperty('_fn');
+    expect(test).toHaveProperty('test');
+    expect(test).not.toHaveProperty('settings');
+    expect(test).not.toHaveProperty('cache');
+    expect(test).not.toHaveProperty('_locals');
   });
 
-  test('Fn: buildTemplate: Should construct a new Template{object}', () => {
-    const result = buildTemplate('', {}, mockContent);
-    //
-    expect(result instanceof Template).toBeTruthy();
-    expect(result.compile instanceof Function).toBeTruthy();
-    expect(result.render()).toBe(mockContent);
+
+  test('Fn: buildTemplate', () => {
+    const test = buildTemplate('', {}, mockString);
+    expect(test instanceof Template).toBeTruthy();
+    expect(test.compile instanceof Function).toBeTruthy();
+    expect(test.render()).toBe(mockString);
   });
 
-  test('Fn: getFileString: Should get context as string from the template file', async () => {
-    const result = [getFileString('test', true), await getFileString('test').then(() => 0)];
-    //
-    expect(result[0]).toEqual(mockContent);
-    expect(result[1]).toEqual(0);
+
+  test('Fn: getTemplate', async () => {
+    const test = await Promise.all([getTemplate('test', {}, true), getTemplate('test', {}).then(template => template)]);
+    expect(test[0] instanceof Template).toBeTruthy();
+    expect(test[1] instanceof Template).toBeTruthy();
   });
 
-  test('Fn: render: Should return the requested content in HTML', async () => {
-    const result = await (render('', mockLocals, (err, context) => {
-      if (err) throw err;
-      return context;
-    })).then(string => {
-      return typeof string === 'string';
-    });
-    //
-    expect(result).toBeTruthy();
+
+  test('Fn: render', async () => {
+    const test = await (render('', mockLocals, mockCallback));
+    expect(test).toEqual(mockString);
   });
 });